fix: circular dependency in imports
