+++ adapters/network.js
@@ -266,6 +266,59 @@ const proxyRequest = async request => {
   return response;
 };
 
+/**
+ * handles http request and sends the response in a simple format that is followed in transformer
+ *
+ * @param {string} requestType - http request type like post, get etc,.
+ * @param {Array<any>} httpArgs - Array of arguments that should be sent to the request
+ * @returns {{httpResponse: any, processedResponse: any}}
+ *  - __httpResponse__: indicates the response we get from httpGET or httpPOST methods
+ *  - __processedResponse__: indicates a wrapeed response object returned from processedAxiosResponse
+ * @example
+ *  handleHttpRequest("post", ["https://example.com", {}])
+ *  handleHttpRequest("constructor", [
+ *    {
+        method: "post",
+        url: "https://myapi.com/api/2/deletions/users",
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: `Basic sdineciuneu839eu3i9/=`
+        },
+        data: {
+          // some random data
+          "key": 1,
+          "ke": 2
+        }
+      }
+    ])
+ */
+const handleHttpRequest = async (requestType = "post", httpArgs) => {
+  let httpWrapperMethod;
+  switch (requestType.toLowerCase()) {
+    case "get":
+      httpWrapperMethod = httpGET;
+      break;
+    case "put":
+      httpWrapperMethod = httpPUT;
+      break;
+    case "patch":
+      httpWrapperMethod = httpPATCH;
+      break;
+    case "delete":
+      httpWrapperMethod = httpDELETE;
+      break;
+    case "constructor":
+      httpWrapperMethod = httpSend;
+      break;
+    default:
+      httpWrapperMethod = httpPOST;
+      break;
+  }
+  const httpResponse = await httpWrapperMethod(...httpArgs);
+  const processedResponse = processAxiosResponse(httpResponse);
+  return { httpResponse, processedResponse };
+};
+
 module.exports = {
   httpSend,
   httpGET,
@@ -276,5 +329,6 @@ module.exports = {
   proxyRequest,
   prepareProxyRequest,
   getPayloadData,
-  getFormData
+  getFormData,
+  handleHttpRequest
 };
